Fixed differentialDrive, commented code
