+++ src/main/java/frc/robot/commands/ArcadeDrive.java
@@ -1,0 +1,46 @@
+package frc.robot.commands;
+
+import edu.wpi.first.wpilibj.GenericHID.Hand;
+import edu.wpi.first.wpilibj.command.Command;
+import frc.robot.Robot;
+
+
+public class ArcadeDrive extends Command {
+
+    public ArcadeDrive() {
+        //requires is method that tells commands which subsystems will be using it
+        //IT IS MANDATORY
+        requires(Robot.driveBase);
+    }
+
+
+    @Override
+    protected void initialize() {
+
+    }
+
+
+    @Override
+    protected void execute() {
+        Robot.driveBase.drive(Robot.oi.controller.getY(Hand.kLeft),//Y-Axis of left joystick
+                              Robot.oi.controller.getX(Hand.kLeft));//X-Axis of right joystick
+    }
+
+    @Override
+    protected boolean isFinished() {
+        // TODO: Make this return true when this Command no longer needs to run execute()
+        return false;
+    }
+
+
+    @Override
+    protected void end() {
+
+    }
+
+
+    @Override
+    protected void interrupted() {
+        super.interrupted();
+    }
+}
